test: put useBreakpoint SSR tests in its own file
